+++ packages/brick_sqlite_abstract/lib/src/db/schema.dart
@@ -1,5 +1,6 @@
 // Heavily, heavily inspired by [Aqueduct](https://github.com/stablekernel/aqueduct/blob/master/aqueduct/lib/src/db/schema/schema_builder.dart)
 // Unfortunately, some key differences such as inability to use mirrors and the sqlite vs postgres capabilities make DIY a more palatable option than retrofitting
+import 'package:brick_sqlite_abstract/src/db/schema/schema_index.dart';
 import 'package:meta/meta.dart' show required, visibleForTesting;
 
 import 'migration.dart';
@@ -120,6 +121,13 @@ class Schema {
         onDeleteCascade: command.onDeleteCascade,
         onDeleteSetDefault: command.onDeleteSetDefault,
       ));
+    } else if (command is CreateIndex) {
+      final table = findTable(command.onTable);
+      table.indices.add(SchemaIndex(
+        columns: command.columns,
+        tableName: command.onTable,
+        unique: command.unique,
+      ));
     } else {
       throw FallThroughError();
     }
@@ -136,9 +144,9 @@ class Schema {
     return '''Schema(
 \t$version,
 \tgeneratorVersion: $generatorVersion,
-\ttables: Set<SchemaTable>.from([
+\ttables: <SchemaTable>{
 \t\t$tableString
-\t])
+\t}
 )'''
         .replaceAll('\t', '  ');
   }