+++ packages/brick_sqlite_abstract/CHANGELOG.md
@@ -1,6 +1,8 @@
 ## Unreleased
 
-* When joins tables are created, add a unique index by both columns to ensure no duplicates are inserted
+## 0.0.8
+
+* When joins tables are created, add a unique index by both columns to ensure no duplicates are inserted. Prior applications will need to manually specify `CreateIndex` in a custom migration for existing joins tables. When new joins tables are generated in migrations, `CreateIndex` will also be generated.
 
 ## 0.0.7+4